Fixes #77
For details see:
http://www.quora.com/Django-web-framework/When-using-Djangos-signals-where-should-the-code-that-registers-a-new-receiver-function-go
